artgram-microservices/services/user-service/requirements.txt
Django==5.0
Pillow==10.2.0
requests==2.31.0
django-cors-headers==4.3.1
python-consul==1.1.0
Flask==3.0.0
